Fix race condition reading/writing log in test
In the test that checks that a cli script logs errors to stderr, there
was a race condition in which the fn would write to the logger at the
same time that the test would read from the buffer in the logger. This
commit adds a buffer type that protects read and write operations with
a sync.RWMutex. It appears this issue only affected the test code.