Stop calling wg.Done twice in cron invoke test
When testing Invoke, there is a potential race condition in the test
between a second invocation of the function and the test getting the
"done" message from the wait group. If the function is invoked again, it
may call Done on the wait group even though there are no resources to
return. This is a bug in the design of the test.

This commit updates the test to ensure that wg.Done is called at most
once during the test.