Remove exename from cli tests
The exename function did not have the intended effect on the tests. When
you provide a -test.run value to go test, you do not need to provide a
.exe extension for Windows. The result of running a test without a
correct test name is "PASS" being written to stdout. This meant that the
tests invoking the cli fn would not receive the expected output. This
commit removes the exename function and its use in the cli tests.