Reduce number of invocations of cli fns in tests
The tests were running a large number of iterations (1000 for a service
fn and 100 for a script fn), but these large numbers are unnecessary and
sometimes caused the tests to fail due to failing to complete within the
allowed 5s timeout. This commit reduces the number of iterations to 10
for each test, which are significantly more achievable targets without
sacrificing the quality of the test.