Add restart options to fnrunner
This commit adds the restart and restart-wait CLI options to fnrunner.
The `restart` option defaults to true and the `restart-wait` option
defaults to 10 seconds. These options are used to restart a running
system if it fails (e.g., due to network errors).

fnrunner will panic when it receives an error running a system and
`restart` is set to false.