Return error when specifying az msg disposition fails
Renewing message locks with the Azure ServiceBus library will fail after
idling for approximately 2 hours due to a bug in the library. This
commit provides a workaround by returning an error from the Serve
function when setting the message disposition fails. This may occur when
a call to RenewLocks fails and the internal context is canceled.

If the runner restarts the system, it will establish a new connection
and should run as expected.